Update Illumina tasks to use prokka; add REPLACE_FASTA parameter
The former is b/c the ilm: namespace still contained the old methods for
annotating assemblies. These have been updated to use the new method
(prokka), and the former steps were deprecated to the old: namespace.

The REPLACE_FASTA parameter allows the polished_assembly.fasta file to
be replaced after the pull_down_raw_reads step, which may be useful in
the event that you need to re-run the pipeline on a manually fixed
assembly.